+++ src/tests/error.test.ts
@@ -24,12 +24,31 @@ test("getting codes works", (t) => {
     const error    = new Error("A bug");
     const sendable = SendableError.of(error);
     t.is(sendable.getMessage(), "A bug");
-    //t.is(sendable.getPublicMessage(), "An internal error occurred");
     t.deepEqual(sendable.toResponse(), {
+      code   : "misc/internal-error",
+      details: {},
+      message: "An internal error occurred",
+      traceId: sendable.getTraceId(),
+    });
+
+    t.deepEqual(SendableError.of(error, {public: true}).toResponse(), {
       code   : "misc/internal-error",
       details: {},
       message: "A bug",
       traceId: sendable.getTraceId(),
     });
+
+    const code = new ErrorCode({
+      id            : "test/error-code",
+      defaultMessage: "Test error code",
+    });
+    t.deepEqual(SendableError.of(error, {public: true, code: code}).toResponse(), {
+      code   : "test/error-code",
+      details: {},
+      message: "Test error code",
+      traceId: sendable.getTraceId(),
+    });
+
+    t.deepEqual(sendable.toResponse(), SendableError.of(sendable).toResponse());
   }
 });